requirements.txt
allure-pytest==2.12.0
pytest==7.2.0
playwright==1.28.0
pytest-playwright==0.3.0